:art: use bigint

+++ src/i18n/locale.rs
@@ -21,11 +21,11 @@ lazy_static! {
 #[derive(Queryable, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Item {
-    pub id: i32,
+    pub id: i64,
     pub lang: String,
     pub code: String,
     pub message: String,
-    pub version: i32,
+    pub version: i64,
     pub created_at: NaiveDateTime,
     pub updated_at: NaiveDateTime,
 }
@@ -50,11 +50,11 @@ pub trait Dao {
     fn count(&self, lang: &str) -> Result<i64>;
     fn all(&self) -> Result<Vec<Item>>;
     fn by_lang(&self, lang: &str) -> Result<Vec<Item>>;
-    fn by_id(&self, id: i32) -> Result<Item>;
+    fn by_id(&self, id: i64) -> Result<Item>;
     fn by_lang_and_code(&self, lang: &str, code: &str) -> Result<Item>;
-    fn delete(&self, id: i32) -> Result<()>;
+    fn delete(&self, id: i64) -> Result<()>;
     fn create(&self, lang: &str, code: &str, message: &str) -> Result<()>;
-    fn update(&self, id: i32, code: &str, message: &str) -> Result<()>;
+    fn update(&self, id: i64, code: &str, message: &str) -> Result<()>;
 }
 
 fn loop_yaml(
@@ -171,7 +171,7 @@ impl Dao for Connection {
             .load::<Item>(self)?;
         Ok(items)
     }
-    fn by_id(&self, id: i32) -> Result<Item> {
+    fn by_id(&self, id: i64) -> Result<Item> {
         let it = locales::dsl::locales
             .filter(locales::dsl::id.eq(id))
             .first::<Item>(self)?;
@@ -184,7 +184,7 @@ impl Dao for Connection {
             .first::<Item>(self)?;
         Ok(it)
     }
-    fn update(&self, id: i32, code: &str, message: &str) -> Result<()> {
+    fn update(&self, id: i64, code: &str, message: &str) -> Result<()> {
         let now = Utc::now().naive_utc();
         let it = locales::dsl::locales.filter(locales::dsl::id.eq(id));
         update(it)
@@ -208,7 +208,7 @@ impl Dao for Connection {
             .execute(self)?;
         Ok(())
     }
-    fn delete(&self, id: i32) -> Result<()> {
+    fn delete(&self, id: i64) -> Result<()> {
         delete(locales::dsl::locales.filter(locales::dsl::id.eq(id))).execute(self)?;
         Ok(())
     }
